+++ src/modules/rkhunter.rs
@@ -70,79 +70,77 @@ impl Module for RkHunter {
         csleep(500).await;
         print("Starting system checks...\r\n").await;
         newline().await;
+        csleep(500).await;
 
-        loop {
-            let task = RKHUNTER_TASKS_LIST.iter().choose(&mut rng).unwrap();
-            print(format!("{task}\r\n")).await;
+        let task = RKHUNTER_TASKS_LIST.iter().choose(&mut rng).unwrap();
+        print(format!("{task}\r\n")).await;
 
-            let is_rootkit = rng.random_bool(0.5);
-            let rk_pad = if is_rootkit { "  " } else { "" };
-            let rootkit = RKHUNTER_ROOTKITS_LIST.iter().choose(&mut rng).unwrap();
-            if is_rootkit {
-                print(format!("  Checking for {rootkit}...\r\n")).await;
+        let is_rootkit = rng.random_bool(0.5);
+        let rk_pad = if is_rootkit { "  " } else { "" };
+        let rootkit = RKHUNTER_ROOTKITS_LIST.iter().choose(&mut rng).unwrap();
+        if is_rootkit {
+            print(format!("  Checking for {rootkit}...\r\n")).await;
+        }
+
+        let mut rootkit_found = false;
+        let num_checks = rng.random_range(2..30);
+        let mut checks: Vec<&&str> = RKHUNTER_CHECKS_LIST.sample(&mut rng, num_checks).collect();
+
+        checks.sort();
+
+        // Calculate the right padding for checks to properly align the check status, with a
+        // minimum width of 40 characters
+        let mut check_pad = 40;
+        for &check in &checks {
+            check_pad = if check.len() > check_pad {
+                check.len()
+            } else {
+                check_pad
+            }
+        }
+
+        for &check in &checks {
+            csleep(rng.random_range(200..1000)).await;
+            // Specify if a check was positive; if yes also set the rootkit to have been found
+            let check_positive = rng.random_bool(check_positive_probability);
+            if check_positive {
+                rootkit_found = true;
             }
 
-            let mut rootkit_found = false;
-            let num_checks = rng.random_range(2..30);
-            let mut checks: Vec<&&str> =
-                RKHUNTER_CHECKS_LIST.sample(&mut rng, num_checks).collect();
-
-            checks.sort();
-
-            // Calculate the right padding for checks to properly align the check status, with a
-            // minimum width of 40 characters
-            let mut check_pad = 40;
-            for &check in &checks {
-                check_pad = if check.len() > check_pad {
-                    check.len()
-                } else {
-                    check_pad
-                }
+            // Prepare check and status
+            let mut check_status = if check_positive {
+                "Found".red()
+            } else {
+                "Not found".resetting()
+            };
+            if rng.random_bool(0.01) {
+                check_status = "Skipped".resetting();
             }
 
-            for &check in &checks {
-                csleep(rng.random_range(200..1000)).await;
-                // Specify if a check was positive; if yes also set the rootkit to have been found
-                let check_positive = rng.random_bool(check_positive_probability);
-                if check_positive {
-                    rootkit_found = true;
-                }
-
-                // Prepare check and status
-                let mut check_status = if check_positive {
-                    "Found".red()
-                } else {
-                    "Not found".resetting()
-                };
-                if rng.random_bool(0.01) {
-                    check_status = "Skipped".resetting();
-                }
-
-                print(format!(
-                    "{rk_pad}  {check:<check_pad$} [ {check_status} ]\r\n"
-                ))
-                .await;
-            }
-
-            if is_rootkit {
-                check_pad += 2;
-                print(format!(
-                    "  {rootkit:<check_pad$} [ {status} ]\r\n",
-                    status = if rootkit_found {
-                        "Found".red()
-                    } else {
-                        "Not found".resetting()
-                    }
-                ))
-                .await;
-            }
-
-            newline().await;
-            csleep(500).await;
+            print(format!(
+                "{rk_pad}  {check:<check_pad$} [ {check_status} ]\r\n"
+            ))
+            .await;
 
             if appconfig.should_exit() {
                 return;
             }
         }
+
+        if is_rootkit {
+            check_pad += 2;
+            print(format!(
+                "  {rootkit:<check_pad$} [ {status} ]\r\n",
+                status = if rootkit_found {
+                    "Found".red()
+                } else {
+                    "Not found".resetting()
+                }
+            ))
+            .await;
+        }
+
+        newline().await;
+        csleep(500).await;
     }
 }